Refactor - Simplify code by moving liquid doc visiting logic to  its own module ---- We didn't need to use the make pattern at least at this stage, so I'm removing the code in ThemeCheckCommon and placing it in its own module.
We may need to revisit this when we are writing theme checks. However, for the purpose of hover, this is where the code should live.
