+++ packages/theme-check-common/src/context-utils.spec.ts
@@ -1,14 +1,11 @@
 import { beforeEach, describe, expect, it } from 'vitest';
-import { MockFileSystem } from './test';
 import {
   makeGetDefaultLocale,
   makeGetDefaultTranslations,
-  makeGetLiquidDocDefinitions,
   makeGetMetafieldDefinitions,
 } from './context-utils';
+import { MockFileSystem } from './test';
 import { AbstractFileSystem } from './AbstractFileSystem';
-import { toSourceCode } from './to-source-code';
-import { LiquidHtmlNode } from '@shopify/liquid-html-parser';
 
 describe('Unit: getDefaultLocale', () => {
   let fs: AbstractFileSystem;
@@ -102,67 +99,6 @@ describe('Unit: getDefaultLocale', () => {
 
       getDefaultTranslations = makeGetDefaultTranslations(fs, [], 'shopify-vfs:/frenchDefault');
       expect(await getDefaultTranslations()).to.eql({ beverage: 'café' });
-    });
-  });
-
-  describe('Unit: makeGetLiquidDocDefinitions', () => {
-    function toAST(code: string) {
-      return toSourceCode('/tmp/foo.liquid', code).ast as LiquidHtmlNode;
-    }
-
-    it('should return name if no valid annotations are present in definition', async () => {
-      const ast = toAST(`
-        {% doc %}
-          just a description
-          @undefined asdf
-        {% enddoc %}
-      `);
-
-      const getLiquidDocDefinitions = makeGetLiquidDocDefinitions();
-      const result = await getLiquidDocDefinitions(ast, 'product-card');
-      expect(result).to.deep.equal({
-        name: 'product-card',
-        parameters: [],
-      });
-    });
-
-    it('should extract name, description and type from param annotations', async () => {
-      const ast = toAST(`
-        {% doc %}
-          @param {String} firstParam - The first param
-          @param {Number} secondParam - The second param
-          @param paramWithNoType - param with no type
-          @param paramWithOnlyName
-        {% enddoc %}
-      `);
-
-      const getLiquidDocDefinitions = makeGetLiquidDocDefinitions();
-      const result = await getLiquidDocDefinitions(ast, 'product-card');
-      expect(result).to.deep.equal({
-        name: 'product-card',
-        parameters: [
-          {
-            name: 'firstParam',
-            description: 'The first param',
-            type: 'String',
-          },
-          {
-            name: 'secondParam',
-            description: 'The second param',
-            type: 'Number',
-          },
-          {
-            name: 'paramWithNoType',
-            description: 'param with no type',
-            type: null,
-          },
-          {
-            name: 'paramWithOnlyName',
-            description: '',
-            type: null,
-          },
-        ],
-      });
     });
   });
 });
